tools/local-training/requirements.txt
accelerate>=0.34.0
datasets>=2.21.0
pillow>=10.4.0
sentencepiece>=0.2.0
torch>=2.4.0
transformers>=4.44.0
